+++ c++/src/MemoryPool.cc
@@ -17,9 +17,9 @@
  */
 
 #include "orc/Int128.hh"
+#include "orc/MemoryPool.hh"
 
 #include "Adaptor.hh"
-#include "MemoryPool.hh"
 
 #include <cstdlib>
 #include <iostream>
@@ -236,70 +236,6 @@ namespace orc {
   #ifdef __clang__
     #pragma clang diagnostic ignored "-Wexit-time-destructors"
   #endif
-
-  BlockBuffer::BlockBuffer(MemoryPool& pool, uint64_t _blockSize)
-      : memoryPool(pool),
-        currentSize(0),
-        currentCapacity(0),
-        blockSize(_blockSize) {
-    if (blockSize == 0) {
-      throw std::logic_error("Block size cannot be zero");
-    }
-    reserve(blockSize);
-  }
-
-  BlockBuffer::~BlockBuffer() {
-    for (size_t i = 0; i < blocks.size(); ++i) {
-      memoryPool.free(blocks[i]);
-    }
-    blocks.clear();
-    currentSize = currentCapacity = 0;
-  }
-
-  uint64_t BlockBuffer::getBlockNumber() const {
-    return (currentSize + blockSize - 1) / blockSize;
-  }
-
-  Block BlockBuffer::getBlock(uint64_t blockIndex) {
-    if (blockIndex >= getBlockNumber()) {
-      throw std::out_of_range("Block index out of range");
-    }
-    return Block(blocks[blockIndex],
-                 std::min(currentSize - blockIndex * blockSize, blockSize));
-  }
-
-  Block BlockBuffer::getEmptyBlock() {
-    if (currentSize < currentCapacity) {
-      Block emptyBlock(blocks[currentSize / blockSize] + currentSize % blockSize,
-          blockSize - currentSize % blockSize);
-      currentSize = (currentSize / blockSize + 1) * blockSize;
-      return emptyBlock;
-    } else {
-      resize(currentSize + blockSize);
-      return Block(blocks.back(), blockSize);
-    }
-  }
-
-  void BlockBuffer::reserve(uint64_t capacity) {
-    while (currentCapacity < capacity) {
-      char* newBlockPtr = memoryPool.malloc(blockSize);
-      if (newBlockPtr != nullptr) {
-        blocks.push_back(newBlockPtr);
-        currentCapacity += blockSize;
-      } else {
-        break;
-      }
-    }
-  }
-
-  void BlockBuffer::resize(uint64_t size) {
-    reserve(size);
-    if (currentCapacity >= size) {
-      currentSize = size;
-    } else {
-      throw std::logic_error("Block buffer resize error");
-    }
-  }
 
   MemoryPool* getDefaultPool() {
     static MemoryPoolImpl internal;
