+++ c++/test/TestBlockBuffer.cc
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-#include "MemoryPool.hh"
+#include "BlockBuffer.hh"
 #include "orc/OrcFile.hh"
 #include "wrap/gtest-wrapper.h"
